+++ packages/antd/src/widgets/CheckboxesWidget/index.tsx
@@ -4,9 +4,7 @@ import {
   ariaDescribedByIds,
   enumOptionsIndexForValue,
   enumOptionsValueForIndex,
-  getTemplate,
   optionId,
-  titleId,
   FormContextType,
   WidgetProps,
   RJSFSchema,
@@ -23,25 +21,8 @@ export default function CheckboxesWidget<
   T = any,
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any
->({
-  autofocus,
-  disabled,
-  formContext,
-  id,
-  label,
-  hideLabel,
-  onBlur,
-  onChange,
-  onFocus,
-  options,
-  readonly,
-  registry,
-  schema,
-  uiSchema,
-  value,
-}: WidgetProps<T, S, F>) {
+>({ autofocus, disabled, formContext, id, onBlur, onChange, onFocus, options, readonly, value }: WidgetProps<T, S, F>) {
   const { readonlyAsDisabled = true } = formContext as GenericObjectType;
-  const TitleFieldTemplate = getTemplate<'TitleFieldTemplate', T, S, F>('TitleFieldTemplate', registry, options);
 
   const { enumOptions, enumDisabled, inline, emptyValue } = options;
 
@@ -65,17 +46,6 @@ export default function CheckboxesWidget<
 
   return Array.isArray(enumOptions) && enumOptions.length > 0 ? (
     <>
-      {!hideLabel && !!label && (
-        <div>
-          <TitleFieldTemplate
-            id={titleId<T>(id)}
-            title={label}
-            schema={schema}
-            uiSchema={uiSchema}
-            registry={registry}
-          />
-        </div>
-      )}
       <Checkbox.Group
         disabled={disabled || (readonlyAsDisabled && readonly)}
         name={id}